Update __init__.py
Had an issue when demandimport was used with a sys.meta_path import hook https://gist.github.com/fried/4254e881ba90cf0d9e4f26f7da3bcceb
 
The hook allows for the following module:  (We have reasons for this, its not as simple as just forcing everyone to use one way or the other)
 package.lib.module
to be imported with the name
 package.module
but maintain that they are the same module object.    
   
Well when testing our hook with demandimport,  demandimport in python 2 was sending imports to our hook in the form  'package.lib.package' after we attempted to demand import 'package.lib.module'.      In python3 it worked as expected. 

I made the following change, and our tests passed and demain imports tests passed.   I know the level field controls absolute/relative imports in python2, and this might break some use case but i don't know what that is.